Split/fuse `xlocinfo.h` to fix unintentional export in `format.cpp` (#1927)
* Extract `__msvc_xlocinfo_types.hpp`.

* Include `__msvc_xlocinfo_types.hpp` instead of `xlocinfo.h`.

In `xwctomb.cpp`, need to reorder functions.

Also, `<cwchar>` provides `wmemcmp`, not `<cstring>`.

* Fuse `xlocinfo.h`.
