+++ stl/inc/header-units.json
@@ -7,6 +7,7 @@
         // "__msvc_all_public_headers.hpp", // for testing, not production
         "__msvc_system_error_abi.hpp",
         "__msvc_tzdb.hpp",
+        "__msvc_xlocinfo_types.hpp",
         "algorithm",
         "any",
         "array",
@@ -130,7 +131,6 @@
         "xlocale",
         "xlocbuf",
         "xlocinfo",
-        "xlocinfo.h",
         "xlocmes",
         "xlocmon",
         "xlocnum",
