+++ stl/src/_tolower.cpp
@@ -5,11 +5,11 @@
 
 #include <yvals.h>
 
+#include <__msvc_xlocinfo_types.hpp>
 #include <cctype>
 #include <clocale>
 #include <cstdlib>
 #include <internal_shared.h>
-#include <xlocinfo.h>
 
 #include "awint.hpp"
