+++ stl/src/xtowupper.cpp
@@ -3,8 +3,8 @@
 
 // _Towupper -- convert wchar_t to upper case
 
+#include <__msvc_xlocinfo_types.hpp>
 #include <cwchar>
-#include <xlocinfo.h>
 
 #include "awint.hpp"
