+++ stl/src/xwcsxfrm.cpp
@@ -3,12 +3,12 @@
 
 // Transform a wide-character string using the locale information as set by LC_COLLATE.
 
+#include <__msvc_xlocinfo_types.hpp>
 #include <climits>
 #include <clocale>
 #include <cstdlib>
 #include <cstring>
 #include <internal_shared.h>
-#include <xlocinfo.h> // for _Collvec, _Wcsxfrm
 
 #include "awint.hpp"
